Re-ran Go source conversion for v1.15.2

+++ src/go-src-converted/cmd/compile/internal/ppc64/opt.cs
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// package ppc64 -- go2cs converted at 2020 August 29 09:25:00 UTC
+// package ppc64 -- go2cs converted at 2020 October 08 04:27:30 UTC
 // import "cmd/compile/internal/ppc64" ==> using ppc64 = go.cmd.compile.@internal.ppc64_package
 // Original source: C:\Go\src\cmd\compile\internal\ppc64\opt.go
 
@@ -17,7 +17,7 @@ namespace @internal
     {
         // Many Power ISA arithmetic and logical instructions come in four
         // standard variants. These bits let us map between variants.
-        public static readonly long V_CC = 1L << (int)(0L); // xCC (affect CR field 0 flags)
-        public static readonly long V_V = 1L << (int)(1L); // xV (affect SO and OV flags)
+        public static readonly long V_CC = (long)1L << (int)(0L); // xCC (affect CR field 0 flags)
+        public static readonly long V_V = (long)1L << (int)(1L); // xV (affect SO and OV flags)
     }
 }}}}
