+++ src/go-src-converted/cmd/link/internal/arm/l.cs
@@ -1,5 +1,5 @@
 // Inferno utils/5l/asm.c
-// https://bitbucket.org/inferno-os/inferno-os/src/default/utils/5l/asm.c
+// https://bitbucket.org/inferno-os/inferno-os/src/master/utils/5l/asm.c
 //
 //    Copyright © 1994-1999 Lucent Technologies Inc.  All rights reserved.
 //    Portions Copyright © 1995-1997 C H Forsyth (forsyth@terzarima.net)
@@ -28,7 +28,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// package arm -- go2cs converted at 2020 August 29 10:04:42 UTC
+// package arm -- go2cs converted at 2020 October 08 04:39:53 UTC
 // import "cmd/link/internal/arm" ==> using arm = go.cmd.link.@internal.arm_package
 // Original source: C:\Go\src\cmd\link\internal\arm\l.go
 
@@ -44,7 +44,7 @@ namespace @internal
         // Writing object files.
 
         // Inferno utils/5l/l.h
-        // https://bitbucket.org/inferno-os/inferno-os/src/default/utils/5l/l.h
+        // https://bitbucket.org/inferno-os/inferno-os/src/master/utils/5l/l.h
         //
         //    Copyright © 1994-1999 Lucent Technologies Inc.  All rights reserved.
         //    Portions Copyright © 1995-1997 C H Forsyth (forsyth@terzarima.net)
@@ -72,12 +72,13 @@ namespace @internal
         // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
         // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
         // THE SOFTWARE.
-        private static readonly long maxAlign = 8L; // max data alignment
-        private static readonly long minAlign = 1L; // min data alignment
-        private static readonly long funcAlign = 4L; // single-instruction alignment
+        private static readonly long maxAlign = (long)8L; // max data alignment
+        private static readonly long minAlign = (long)1L; // min data alignment
+        private static readonly long funcAlign = (long)4L; // single-instruction alignment
 
         /* Used by ../internal/ld/dwarf.go */
-        private static readonly long dwarfRegSP = 13L;
-        private static readonly long dwarfRegLR = 14L;
+        private static readonly long dwarfRegSP = (long)13L;
+        private static readonly long dwarfRegLR = (long)14L;
+
     }
 }}}}
