+++ src/go-src-converted/math/asinh.cs
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// package math -- go2cs converted at 2020 August 29 08:44:44 UTC
+// package math -- go2cs converted at 2020 October 08 03:25:09 UTC
 // import "math" ==> using math = go.math_package
 // Original source: C:\Go\src\math\asinh.go
 
@@ -49,20 +49,22 @@ namespace go
 
         private static double asinh(double x)
         {
-            const float Ln2 = 6.93147180559945286227e-01F; // 0x3FE62E42FEFA39EF
-            const float NearZero = 1.0F / (1L << (int)(28L)); // 2**-28
-            const long Large = 1L << (int)(28L); // 2**28 
+            const float Ln2 = (float)6.93147180559945286227e-01F; // 0x3FE62E42FEFA39EF
+            const float NearZero = (float)1.0F / (1L << (int)(28L)); // 2**-28
+            const long Large = (long)1L << (int)(28L); // 2**28 
             // special cases
             if (IsNaN(x) || IsInf(x, 0L))
             {>>MARKER:FUNCTION_Asinh_BLOCK_PREFIX<<
                 return x;
             }
+
             var sign = false;
             if (x < 0L)
             {
                 x = -x;
                 sign = true;
             }
+
             double temp = default;
 
             if (x > Large) 
@@ -77,7 +79,9 @@ namespace go
             {
                 temp = -temp;
             }
+
             return temp;
+
         }
     }
 }
